Separate web interface from phishing detection, mostly
